Changed topic of Proposal
The changed topic of proposal from an NBA Reddit stat bot to a Pokemon Reddit stat bot. This was due to the fact that I was unable to get a good free NBA API. All of the API's that I was looking at either cost a large amount of money ($300) or were not supported anymore (espn). This pokemon API is actively maintained and free to use.